Move cli to a separate folder to avoid clutter in init
